refactor: move EditListForm and EditListMembersForm to form package
