+++ domain/content/repository/src/commonTest/kotlin/com/livefast/eattrash/raccoonforfriendica/domain/content/repository/DefaultCirclesRepositoryTest.kt
@@ -1,9 +1,9 @@
 package com.livefast.eattrash.raccoonforfriendica.domain.content.repository
 
 import com.livefast.eattrash.raccoonforfriendica.core.api.dto.Account
-import com.livefast.eattrash.raccoonforfriendica.core.api.dto.EditListForm
-import com.livefast.eattrash.raccoonforfriendica.core.api.dto.EditListMembersForm
 import com.livefast.eattrash.raccoonforfriendica.core.api.dto.UserList
+import com.livefast.eattrash.raccoonforfriendica.core.api.form.EditListForm
+import com.livefast.eattrash.raccoonforfriendica.core.api.form.EditListMembersForm
 import com.livefast.eattrash.raccoonforfriendica.core.api.provider.ServiceProvider
 import com.livefast.eattrash.raccoonforfriendica.core.api.service.ListService
 import com.livefast.eattrash.raccoonforfriendica.domain.content.data.CircleReplyPolicy
